Completely switched over to service manager
Much better now! Minimal calls on find and objects in service locator will be populated as needed, so there is never an issue of a service not being loaded when a game object needs it. However, the overall structure of the classes could be more hierarchial, so they dont all need to access each other alll the time. I think the turn manager/ gameflow manager/ could be doing more so that the characters dont need to time their interactions in some weird way.

Will continue to refactor, as well as working through list of additional improvements, including adding in wire prefab, and having the game dispay the ai movement choices.